+++ src/modules/gatsby-source-url/createImgixFluidFieldConfig.ts
@@ -19,7 +19,11 @@ import {
   taskEitherFromSourceDataResolver,
 } from '../../common/utils';
 import { IImgixParams, ImgixFluidArgsResolved } from '../../publicTypes';
-import { createImgixFluidType, ImgixParamsInputType } from './graphqlTypes';
+import {
+  createImgixFluidType,
+  ImgixParamsInputType,
+  unTransformParams,
+} from './graphqlTypes';
 import { buildFluidObject } from './objectBuilders';
 import { resolveDimensions } from './resolveDimensions';
 
@@ -86,6 +90,10 @@ export const createImgixFluidFieldConfig = <TSource, TContext>({
     pipe(
       Do(TE.taskEither)
         .let('rootValue', rootValue)
+        .let('modifiedArgs', {
+          ...args,
+          imgixParams: unTransformParams(args.imgixParams),
+        })
         .sequenceSL(({ rootValue }) => ({
           url: resolveUrlFromSourceData(resolveUrl)(rootValue),
           manualWidth: pipe(
@@ -108,10 +116,10 @@ export const createImgixFluidFieldConfig = <TSource, TContext>({
             client: imgixClient,
           }),
         )
-        .return(({ url, dimensions: { width, height } }) =>
+        .return(({ url, modifiedArgs, dimensions: { width, height } }) =>
           buildFluidObject({
             client: imgixClient,
-            args,
+            args: modifiedArgs,
             sourceHeight: height,
             sourceWidth: width,
             url,