+++ test/unit/source-url/source-url.test.ts
@@ -76,6 +76,26 @@ describe('createResolvers', () => {
       },
       assertion: (url) => expect(url).toMatch(/txt=txt/),
     });
+
+    testForEveryFieldSrcAndSrcSet({
+      name: 'integer and float parameters are set correctly',
+      resolveFieldOpts: {
+        fieldParams: {
+          imgixParams: {
+            blendH: 0.5, // type has both unit_scalar and integer
+            blendW: 10, // type has both unit_scalar and integer
+            txtSize: 1, // type is integer only
+            fpY: 0.5, // type is unit_scalar only
+          },
+        },
+      },
+      assertion: (url) => {
+        expect(url).toMatch('blend-h=0.5');
+        expect(url).toMatch('blend-w=10');
+        expect(url).toMatch('txt-size=1');
+        expect(url).toMatch('fp-y=0.5');
+      },
+    });
   });
   describe('url field', () => {
     it('applies imgixParams correctly', async () => {